+++ R/tests.R
@@ -46,27 +46,6 @@ levenesTest <- function(...) car::leveneTest(...)
 
 
 
-
-#' @title Performs a test for outliers
-#'
-#' @description Performs a test for outliers
-#'
-#' @param \dots Arguments to pass through to \code{\link[car]{outlierTest}} in \pkg{car}.
-#'
-#' @return See help for \code{\link[car]{outlierTest}} in \pkg{car}.
-#'
-#' @seealso \code{\link[car]{outlierTest}} in \pkg{car}.
-#' 
-#' @note This is a simple pass-through to  \code{\link[car]{outlierTest}} in \pkg{car}.
-#' 
-#' @keywords misc
-#' 
-#' @export
-outlierTest <- function(...) car::outlierTest(...)
-
-
-
-
 #' @title Z test for known population standard deviation
 #' 
 #' @description Compute the test of hypothesis and compute confidence interval on the mean of a population when the standard deviation of the population is known.
@@ -99,3 +78,11 @@ outlierTest <- function(...) car::outlierTest(...)
 #' @importFrom TeachingDemos z.test
 #' @export
 NULL
+
+#' @importFrom mosaic t.test
+#' @export
+mosaic::t.test
+
+#' @importFrom car outlierTest
+#' @export
+car::outlierTest